make: Optionally include debug symbols with builds
Per linkerd/linkerd2#2199, we want to be able to include the proxy's
debug symbols in auxiliary diagnostic containers. In order to do this,
the process that produces the proxy's binaries also should be responsible
for producing its debug symbols.

This change adds a new environment variable, `CARGO_DEBUG`, which
the Makefile uses to instrument the build to produce debug symbols. If
`objcopy` is available on the local system (i.e. on Linux), then the
symbols are copied from the binary into a .obj file, and the executable
is stripped from 3.3G down to 9.9M before packaging.

This increases the size of the compressed package artifact by 76% (from 5.0M to
8.8M). This may have similar impacts on build time.